moved callback() to OnlyofficeCallbackController
