+++ src/Controller/OnlyofficeEditorController.php
@@ -4,12 +4,10 @@ namespace Drupal\onlyoffice_connector\Controller;
 
 use Drupal\Core\Url;
 use Drupal\Core\Controller\ControllerBase;
-use Drupal\Core\File\FileSystemInterface;
 use Drupal\Core\Render\RendererInterface;
 use Drupal\file\Entity\File;
 use Drupal\media\Entity\Media;
 use Drupal\onlyoffice_connector\OnlyofficeDocumentHelper;
-use Drupal\user\UserStorageInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
@@ -19,59 +17,46 @@ use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 /**
  * Returns responses for ONLYOFFICE Connector routes.
  */
-class OnlyofficeEditorController extends ControllerBase
-{
+class OnlyofficeEditorController extends ControllerBase {
 
   /**
+   * The renderer service.
+   *
    * @var \Drupal\Core\Render\RendererInterface
    */
   protected $renderer;
 
   /**
+   * The onlyoffice document helper service.
+   *
    * @var \Drupal\onlyoffice_connector\OnlyofficeDocumentHelper
    */
-  protected $docHelper;
+  protected $documentHelper;
 
   /**
-   * The user storage.
-   *
-   * @var \Drupal\user\UserStorageInterface
-   */
-  protected $userStorage;
-
-  /**
+   * Constructs an OnlyofficeEditorController object.
    *
    * @param \Drupal\Core\Render\RendererInterface $renderer
+   * The renderer service.
+   * @param \Drupal\onlyoffice_connector\OnlyofficeDocumentHelper $document_helper
+   * The onlyoffice document helper service.
    */
-  public function __construct(OnlyofficeDocumentHelper $docHelper, UserStorageInterface $userStorage, RendererInterface $renderer)
-  {
-    $this->docHelper = $docHelper;
+  public function __construct(RendererInterface $renderer, OnlyofficeDocumentHelper $document_helper) {
     $this->renderer = $renderer;
-    $this->userStorage = $userStorage;
+    $this->documentHelper = $document_helper;
   }
 
-  public static function create(ContainerInterface $container)
-  {
+  /**
+   * {@inheritdoc}
+   */
+  public static function create(ContainerInterface $container) {
     return new static(
+      $container->get('renderer'),
       $container->get('onlyoffice_connector.document_helper'),
-      $container->get('entity_type.manager')->getStorage('user'),
-      $container->get('renderer')
     );
   }
 
-  public function editor(Media $media, Request $request)
-  {
-    if ($request->isMethod('get')) {
-      return $this->editGet($media, $request);
-    } else if ($request->isMethod('post')) {
-      return $this->editPost($media, $request);
-    } else {
-      throw new BadRequestHttpException();
-    }
-  }
-
-  public function editGet(Media $media, Request $request)
-  {
+  public function editor(Media $media, Request $request) {
     $build = [
       'page' => $this->getDocumentConfig($media)
     ];
@@ -85,69 +70,7 @@ class OnlyofficeEditorController extends ControllerBase
     return $response;
   }
 
-  public function editPost(Media $media, Request $request)
-  {
-    $errorMessage = NULL;
-
-    $body = [];
-    $content = $request->getContent();
-    if (!empty($content)) {
-      $body = json_decode($content, true);
-    }
-
-    // ToDo: check if null etc+
-
-    $account = $this->userStorage->load($body["actions"][0]["userid"]);
-    \Drupal::currentUser()->setAccount($account);
-
-    $status = OnlyofficeEditorController::CALLBACK_STATUS[$body["status"]];
-
-    switch ($status) {
-      case "Editing":
-        // ToDo: check if locking mechanisms exist
-        break;
-      case "MustSave":
-      case "Corrupted":
-        $errorMessage = $this->proccess_save($body, $media);
-        break;
-      case "MustForceSave":
-      case "CorruptedForceSave":
-        break;
-    }
-    //https://api.drupal.org/api/drupal/core%21modules%21file%21file.module/function/file_save_data/8.8.x
-
-    if ($errorMessage == NULL) {
-      return new JsonResponse(['error' => 0], 200);
-    } else {
-      return new JsonResponse(['error' => 1, 'message' => $errorMessage], 400);
-    }
-  }
-
-  private function proccess_save($body, Media $media)
-  {
-    $fid = $media->toArray()["field_media_document"][0]["target_id"];
-    $file = File::load($fid);
-
-    $download_url = $body["url"];
-    if ($download_url === null) {
-      return 'nothing to save';
-    }
-
-    $new_data = file_get_contents($download_url);
-    if ($new_data === null) return 'nothing to save';
-
-    $filepath = $file->getFileUri();
-    $result = \Drupal::service('file_system')->saveData($new_data, $filepath, FileSystemInterface::EXISTS_REPLACE);
-
-    if ($result === 0) return 'saving failed';
-    $file->setSize(strlen($new_data));
-    $file->save();
-
-    return NULL;
-  }
-
-  private function getDocumentConfig(Media $media)
-  {
+  private function getDocumentConfig(Media $media) {
     $fid = $media->toArray()["field_media_document"][0]["target_id"];
     $file = File::load($fid);
 
@@ -156,14 +79,14 @@ class OnlyofficeEditorController extends ControllerBase
     $author = $file->getOwner();
     $user = \Drupal::currentUser()->getAccount();
     $filename = $file->getFilename();
-    $extension = $this->docHelper->getExtension($filename);
+    $extension = $this->documentHelper->getExtension($filename);
 
-    $can_edit = $this->docHelper->isEditable($extension) || $this->docHelper->isFillForms($extension);
+    $can_edit = $this->documentHelper->isEditable($extension) || $this->documentHelper->isFillForms($extension);
     $edit_permission = $media->access("update", $user);
 
     $config = [
       'type' => 'desktop',
-      'documentType' => $this->docHelper->getDocumentType($extension),
+      'documentType' => $this->documentHelper->getDocumentType($extension),
       'document' => [
         'title' => $filename,
         'url' => $file->createFileUrl(false),
@@ -181,7 +104,7 @@ class OnlyofficeEditorController extends ControllerBase
       'editorConfig' => [
         'mode' => $can_edit ? 'edit' : 'view',
         'lang' => 'en', // ToDo: change to user language
-        'callbackUrl' => Url::fromRoute('onlyoffice_connector.editor', ['media' => $media->id()], ['absolute' => true])->toString(),
+        'callbackUrl' => Url::fromRoute('onlyoffice_connector.callback', ['uuid' => $media->uuid()], ['absolute' => true])->toString(),
         'user' => [
           'id' => $user->id(),
           'name' => $user->getDisplayName()
@@ -199,14 +122,4 @@ class OnlyofficeEditorController extends ControllerBase
       '#forms_unavailable_notice' => $this->t('forms_unavailable_notice')
     ];
   }
-
-  const CALLBACK_STATUS = array(
-    0 => 'NotFound',
-    1 => 'Editing',
-    2 => 'MustSave',
-    3 => 'Corrupted',
-    4 => 'Closed',
-    6 => 'MustForceSave',
-    7 => 'CorruptedForceSave'
-  );
 }
